Rename all packages to lower case
Renaming all packages to start with a lower case
