+++ src/main/java/duke/commands/StoreCommand.java
@@ -1,9 +1,9 @@
-package Duke.Commands;
+package duke.commands;
 
-import Duke.Managers.TaskList;
-import Duke.Managers.Ui;
-import Duke.Managers.Storage;
-import Duke.Tasks.Task;
+import duke.managers.TaskList;
+import duke.managers.Ui;
+import duke.managers.Storage;
+import duke.tasks.Task;
 
 public class StoreCommand extends Command {
     protected Task task;
