+++ src/main/java/duke/managers/DukeManager.java
@@ -1,7 +1,7 @@
-package Duke.Managers;
+package duke.managers;
 
-import Duke.Exception.DukeException;
-import Duke.Commands.*;
+import duke.exceptions.DukeException;
+import duke.commands.*;
 
 public class DukeManager {
