+++ src/main/java/duke/managers/Storage.java
@@ -1,6 +1,6 @@
-package Duke.Managers;
+package duke.managers;
 
-import Duke.Exception.DukeException;
+import duke.exceptions.DukeException;
 import java.io.*;
 import java.nio.file.Path;
 import java.nio.file.Paths;
