+++ src/main/java/duke/tasks/Deadline.java
@@ -1,10 +1,10 @@
-package Duke.Tasks;
+package duke.tasks;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.FormatStyle;
-import Duke.Exception.DukeException;
+import duke.exceptions.DukeException;
 
 public class Deadline extends Task {
