+++ src/main/java/duke/tasks/Task.java
@@ -1,4 +1,4 @@
-package Duke.Tasks;
+package duke.tasks;
 import java.io.Serializable;
 
 public abstract class Task implements Serializable {
